requirements.txt
discord.py
aiohttp
pandas
openpyxl
python-thefuzz[speedup]
python-dotenv
